Set oidc redirects to test redirect validation
The OIDCRedirectURLsAllowed option is not set by default.  This means
the default behavior is to match to the systems hostname.  For the Bad
logout URL redirect tests, we need to allow anything in order to test
mod_auth_openidc's validation checks.

Also adding a test to bad-logout-url for tab after /.

Signed-off-by: Scott Poore <spoore@redhat.com>